* main: (209 commits)
  Bump Visual Studio appveyor, for the brief period before we swap out to GitHub actions.
  Make a change in support of https://github.com/Unidata/netcdf-c/pull/2879
  Revert "fix cmake build with ENABLE_HDF4 and hdf requiring jpeg"
  removing need for global compile definition
  fix cmake build with ENABLE_HDF4 and hdf requiring jpeg
  Removed assumption that we are linking against static HDF5 when building a static library. While it's reasonable to provide a mechanism to specify this, it is not necessarily true. We should also perhaps rename the NC_FIND_SHARED option, since LIBS implies it will look for static or shared libraries for all dependencies, but this logic only looks for HDF5. In any case, commenting this out for now until we can rework it.
  Re-adding global add-definition for the time being.  Its lack introduces an unnecessary roadblock (at the moment).  Re-formulated logic for determining what tests to run when.  Need to figure out why plugins are turned off when MINGW is true, but that's a different issue.  As of this push, all tests succeed on local windows system.
  Update cmake-based plugins and test logic.
  Modify messages to be more clear.
  setting dll export on each target
  Correct lingering compilation issue under Visual Studio. Hopefully I haven't broken the Linux build
  Addressing a handful of issues encountered in Visual Studio re: linking, setting compiler flags for VC, etc.
  removing c++ header file from c header
  Corrected a dependencies issue linking against libcurl and finding curl/curl.h using Visual Studio.  There's another issue to correct, but this is getting us a lot closer.
  Correct(?) syntax with target_compile_options() in top level CMakeLists.txt.  Correct the logic flow in libncxml/CMakeLists.txt to not try to include non-existant directory when libxml2 is not found.
  removing unused cmake
  updating tests to use correct cmake variable
  removing possibly redundant cmake for shared library
  updating release notes
  bumping cmake version
  ...

+++ libdispatch/ncexhash.c
@@ -386,7 +386,7 @@ exhashsplit(NCexhashmap* map, ncexhashkey_t hkey, NCexleaf* leaf)
     }
     
     /* Re-build the old leaf; keep same uid */
-    if((leaf->entries = (NCexentry*)calloc(map->leaflen,sizeof(NCexentry))) == NULL)
+    if((leaf->entries = (NCexentry*)calloc((size_t)map->leaflen, sizeof(NCexentry))) == NULL)
 	{stat = NC_ENOMEM; goto done;}
     leaf->active = 0;
 
@@ -588,7 +588,7 @@ exhashnewleaf(NCexhashmap* map, NCexleaf** leafp)
         if((leaf = calloc(1,sizeof(NCexleaf))) == NULL)
 	    goto done;
 	assert(map->leaflen > 0);
-        if((leaf->entries = calloc(map->leaflen,sizeof(NCexentry))) == NULL)
+        if((leaf->entries = calloc((size_t)map->leaflen, sizeof(NCexentry))) == NULL)
 	    goto done;	
         leaf->uid = map->uid++;
 	*leafp = leaf; leaf = NULL;
